// lib/enum.ts
export enum OperationIndex { CREATE, RETRIEVE, UPDATE, DELETE }

export enum OperationName {
    CREATE = "CREATE",
    RETRIEVE = "RETRIEVE",
    UPDATE = "UPDATE",
    DELETE = "DELETE"
}

export enum ModuleId {
    ROLES = "000000000000000000000000",
    VEHICLES = "000000000000000000000001",
    BANK_ACCOUNTS = "000000000000000000000002",
    USERS = "000000000000000000000003",
    HOSTED_TRIPS = "000000000000000000000004",
    REQUESTED_TRIPS = "000000000000000000000005"
}

export enum ModuleName {
    ROLES = "ROLES",
    VEHICLES = "VEHICLES",
    BANK_ACCOUNTS = "BANK_ACCOUNTS",
    USERS = "USERS",
    HOSTED_TRIPS = "HOSTED_TRIPS",
    REQUESTED_TRIPS = "REQUESTED_TRIPS",
}